% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_ecuador_tmax.R
\docType{data}
\name{ecuador.tmax.stations.df}
\alias{ecuador.tmax.stations.df}
\title{Spatial location of meteorological stations in Ecuador}
\format{
A data.frame with eight fields:
*) CODIGO        : identifier of each station
*) NOMBRE        : station name
*) CUENCA_INAMHI : basin name
*) PROVINCIA     : province name
*) CANTON        : canton name
*) PARROQUIA     : parish name
*) LAT           : northing coordinate of the station, EPSG:4326
*) LON           : easting coordinate of the station, EPSG:4326
}
\usage{
data(ecuador.tmax.stations.df)
}
\description{
Spatial location of the 34 stations with daily maximum temperature for Ecuador (dataset ecuador.tmax.zoo).
}
\keyword{datasets}
