% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_ecuador_tmax.R
\docType{data}
\name{ecuador.tmax.wrf.out}
\alias{ecuador.tmax.wrf.out}
\title{WRF (model-based) output for maximum temperature}
\format{
A RasterBrick frame with 10 layers (1 for each day), and a geographical area of 44x47 for a 10Km spatial resolution
}
\usage{
data(ecuador.tmax.wrf.out)
}
\description{
Model-based maximum temperature datasets providing global spatial structure
}
\keyword{datasets}
