% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_ecuador_tmax.R
\docType{data}
\name{ecuador.tmax.zoo}
\alias{ecuador.tmax.zoo}
\title{Ground-based observations of maximum temperature for Ecuador}
\format{
A zoo object with 34 columns (one for each station) and 10 rows (one for each day in jan-2004).
}
\usage{
data(ecuador.tmax.zoo)
}
\description{
A zoo data frame containing time series for january 2004
on 34 meteorological stations located in continental Ecuador.
}
\keyword{datasets}
