% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\docType{data}
\name{elements}
\alias{elements}
\title{elements data}
\format{
A data frame with 8418 rows and 3 columns index1 index2 index3.
}
\usage{
data(elements,"VARMER")
}
\description{
A table of elements
}
\keyword{datasets}
