% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmer.ts.R
\name{idw.uhat.from.sf}
\alias{idw.uhat.from.sf}
\title{IDW interpolation for varmer}
\usage{
idw.uhat.from.sf(stations.sf, reference.raster, idw.formula)
}
\arguments{
\item{stations.sf}{data.frame with metadata for the observations}

\item{reference.raster}{raster image which provides the base structure for the
resulting IDW interpolated image}

\item{idw.formula}{The formula to be passed for the IDW interpolation method}
}
\description{
It produces an IDW-interpolated image from observations
}
