% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U.R
\docType{data}
\name{U}
\alias{U}
\title{U data}
\format{
A data frame with 4320 rows and 3 columns x y and data
}
\usage{
data(U,"VARMER")
}
\description{
A data frame containing the data to make a corrected raster by the Variational Merging method.
}
\keyword{datasets}
