% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_v-0.1.0.R
\docType{data}
\name{uhat.raster}
\alias{uhat.raster}
\title{uhat.raster data}
\format{
A raster with 70 rows and 62 columns with UTM 17S as CRS.
}
\usage{
data(uhat_raster,"VARMER")
}
\description{
A raster layer containing temperature values of interpolated stations
}
\keyword{datasets}
