% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_v-0.1.0.R
\docType{data}
\name{uhat}
\alias{uhat}
\title{uhat data}
\format{
A data frame with 4340 rows and 3 columns x and y are column 1 and 2 and data is the next column.
}
\usage{
data(uhat,"VARMER")
}
\description{
A data frame containing the data present in the raster uhat.raster.
}
\keyword{datasets}
