% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\docType{data}
\name{v}
\alias{v}
\title{v data}
\format{
A data frame with 4340 rows and 3 columns x and y are column 1 and 2 and data is the next column.
}
\usage{
data(v,"VARMER")
}
\description{
A data frame containing the data present in the raster v.raster.
}
\keyword{datasets}
