% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\docType{data}
\name{coordinates}
\alias{coordinates}
\title{coordinates data}
\format{
A data frame with 4340 rows and 3 columns (x y and index).
}
\usage{
data(coordinates,"VARMER")
}
\description{
A data containing the nodes coordinates of uhat.raster and v.raster.
}
\keyword{datasets}
